% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_gene_locations_to_seurat}
\alias{add_gene_locations_to_seurat}
\title{Add gene genomic location to seurat object}
\usage{
add_gene_locations_to_seurat(obj, ref = "hg19")
}
\arguments{
\item{obj}{Seurat object}

\item{ref}{Human genome reference, default `hg19`

export}
}
\description{
Add gene genomic location to seurat object
}
