% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callASCN.R
\name{callAlleleSpecificCNfromHSCN}
\alias{callAlleleSpecificCNfromHSCN}
\title{Call allele specific copy number in single cell datasets}
\usage{
callAlleleSpecificCNfromHSCN(
  hscn,
  eps = 0.000000000001,
  maxCN = NULL,
  selftransitionprob = 0.95,
  progressbar = TRUE,
  ncores = 1,
  fillmissing = TRUE
)
}
\arguments{
\item{hscn}{hscn object from callHaplotypeSpecificCN}

\item{eps}{default 1e-12}

\item{maxCN}{maximum copy number to infer allele specific states, default=NULL which will use the maximum state from CNbins}

\item{selftransitionprob}{probability to stay in the same state in the HMM, default = 0.999, set to 0.0 for an IID model}

\item{progressbar}{Boolean to display progressbar or not, default = TRUE, will only show if ncores == 1}

\item{ncores}{Number of cores to use, default = 1}

\item{fillmissing}{For bins with missing counts fill in values based on neighbouring bins}
}
\value{
allele specific copy number object which includes dataframe similar to input with additional columns which include

* `Maj` (Major allele copy number)
* `Min` (Minor allele copy number)
* `state_AS_phased` (phased state of the form Maj|Min )
* `state_AS` (mirrored state of the form Maj|Min)
* `LOH` (is bin LOH or not)
* `state_phase` (state describing which is the dominant allele and whether it is LOH or not)
* `state_BAF` (binned discretized BAF value calculated as Min / (Maj + Min))
}
\description{
Call allele specific copy number in single cell datasets
}
\details{
In the allele specific copy number inference Maj is always > Min and state_AS_phased == state_AS
}
