% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_one_of}
\alias{check_one_of}
\title{Check that a value is one of the allowed options}
\usage{
check_one_of(x, allowed, name)
}
\arguments{
\item{x}{Value to check}

\item{allowed}{Character vector of allowed values}

\item{name}{Parameter name for error messages}
}
\value{
TRUE if valid, FALSE otherwise (with warning)
}
\description{
Check that a value is one of the allowed options
}
\keyword{internal}
