% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_probability}
\alias{check_probability}
\title{Check that a value is a probability (between 0 and 1)}
\usage{
check_probability(x, name, strict = FALSE)
}
\arguments{
\item{x}{A numeric value to check}

\item{name}{Parameter name for error messages}

\item{strict}{If TRUE, value must be strictly between 0 and 1; if FALSE, endpoints allowed}
}
\value{
TRUE if valid, FALSE otherwise (with warning)
}
\description{
Check that a value is a probability (between 0 and 1)
}
\keyword{internal}
