% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedata.R
\name{computehaplotypecounts}
\alias{computehaplotypecounts}
\title{Compute phasing using top imbalanced cells}
\usage{
computehaplotypecounts(haplotypes, ncells = 10, arm = FALSE)
}
\arguments{
\item{haplotypes}{A data.frame with formatted haplotype data. Required columns:
`chr`, `start`, `end`, `hap_label`, `allele1`, `allele0`, `cell_id`.}

\item{ncells}{Number of top imbalanced cells to use for phasing. Default 10.}

\item{arm}{If TRUE, perform phasing per chromosome arm rather than whole chromosome.
Default FALSE.}
}
\value{
A data.frame with phasing information containing:
  * `chr`, `start`, `end`, `hap_label`: Haplotype block identifiers
  * `phase`: Which allele should be treated as alleleA
}
\description{
Determines the phase for each haplotype block using the top N cells with the
highest allelic imbalance. This method is more robust when there are many
balanced cells that could add noise to phasing.
}
\seealso{
[phase_haplotypes()] for phasing using all cells
}
