% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{createCNmatrix}
\alias{createCNmatrix}
\title{Create copy number matrix for heatmap visualization}
\usage{
createCNmatrix(
  CNbins,
  field = "state",
  maxval = 11,
  na.rm = FALSE,
  fillna = FALSE,
  fillnaplot = FALSE,
  wholegenome = FALSE,
  genome = "hg19",
  centromere = FALSE,
  plotallbins = FALSE
)
}
\arguments{
\item{CNbins}{Copy number data frame with chr, start, end, cell_id columns}

\item{field}{Column to use for values in matrix}

\item{maxval}{Maximum value to cap state at}

\item{na.rm}{Remove rows with NA values}

\item{fillna}{Fill NA values by propagating neighboring values}

\item{fillnaplot}{Fill NA values while preserving centromeric gaps}

\item{wholegenome}{Expand to all genomic bins (deprecated, use plotallbins)}

\item{genome}{Genome assembly ("hg19" or "hg38")}

\item{centromere}{Mark centromeric regions as NA after filling}

\item{plotallbins}{Include all genomic bins, marking centromeres with sentinel value}
}
\description{
Create copy number matrix for heatmap visualization
}
