% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragile_sites_location}
\alias{fragile_sites_location}
\title{Fragile site locations}
\format{
A data frame with 76 rows and 6 columns:
\describe{
\item{seqnames, start, end}{fragile site interval, chromosomes \code{chr}-prefixed}
\item{FSname}{fragile site name}
\item{metadata}{description of the site}
\item{frequency}{\code{common} or \code{rare}}
}
}
\usage{
fragile_sites_location
}
\description{
The underlying fragile site intervals that \link{fragile_sites} bins.
}
\keyword{internal}
