% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_centromere_bins}
\alias{get_centromere_bins}
\title{Get centromeric bins}
\usage{
get_centromere_bins(genome = "hg19", binsize = 5e+05)
}
\arguments{
\item{genome}{Genome assembly, either "hg19" or "hg38". Default is "hg19".}

\item{binsize}{Size of genomic bins in base pairs. Default is 500000 (500kb).}
}
\value{
A data.frame with chr, start, end columns for bins overlapping centromeres.
}
\description{
Identifies genomic bins that overlap with centromeric regions based on cytoband data.
}
