% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getBins}
\alias{getBins}
\title{Make fixed-width bins}
\usage{
getBins(chrom.lengths = hg19_chrlength, binsize = 1000000,
  chromosomes = NULL)
}
\arguments{
\item{chrom.lengths}{A named character vector with chromosome lengths. Names correspond to chromosomes.}

\item{binsize}{Size of bins in basepairs}

\item{chromosomes}{A subset of chromosomes for which the bins are generated.}
}
\value{
A data frame with chr start and end positions
}
\description{
Make fixed-width bins based on given bin size.
}
