% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{hc_clustering}
\alias{hc_clustering}
\title{Hierarchical clustering (deprecated)}
\usage{
hc_clustering()
}
\value{
NULL (with deprecation warning)
}
\description{
This function is deprecated and not implemented. Use [umap_clustering()] instead.
}
