% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_chrlength}
\alias{hg19_chrlength}
\title{hg19 chromosome lengths}
\format{
A named integer vector of length 25, names being chromosome names.
}
\usage{
hg19_chrlength
}
\description{
Chromosome lengths in base pairs, the default used by \code{\link[=getBins]{getBins()}} to build
fixed-width bins.
}
\keyword{internal}
