% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19chrom_coordinates}
\alias{hg19chrom_coordinates}
\title{hg19 chromosome and arm coordinates}
\format{
A data frame with 72 rows and 5 columns: \code{chr}, \code{start}, \code{end}, \code{arm}
and \code{chrarm}.
}
\usage{
hg19chrom_coordinates
}
\description{
Chromosome and chromosome-arm coordinates, used when mapping RNA data onto
arm-level segments.
}
\keyword{internal}
