% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callASCN.R
\name{logspace_add}
\alias{logspace_add}
\title{Add two numbers in log space}
\usage{
logspace_add(logx, logy)
}
\arguments{
\item{logx}{First value in log space.}

\item{logy}{Second value in log space.}
}
\value{
log(exp(logx) + exp(logy))
}
\description{
Computes log(exp(logx) + exp(logy)) in a numerically stable way.
}
\keyword{internal}
