% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{make_ideogram_annotation}
\alias{make_ideogram_annotation}
\title{Create ideogram annotation for heatmap}
\usage{
make_ideogram_annotation(copynumber, genome = "hg19", ideogram_height = 0.3)
}
\arguments{
\item{copynumber}{Copy number matrix with column names in format "chr:start:end"}

\item{genome}{Genome assembly ("hg19" or "hg38"). Default is "hg19".}

\item{ideogram_height}{Height of the ideogram annotation in cm. Default is 0.3.}
}
\value{
A ComplexHeatmap HeatmapAnnotation object
}
\description{
Creates a ComplexHeatmap annotation showing chromosome ideogram (cytobands)
at the bottom of the heatmap.
}
\keyword{internal}
