% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCNprofileBAF}
\alias{plotCNprofileBAF}
\title{Plot a single cell allele specific copy number profile}
\usage{
plotCNprofileBAF(
  cn,
  cellid = NULL,
  chrfilt = NULL,
  pointsize = 1,
  alphaval = 0.6,
  maxCN = 10,
  cellidx = 1,
  BAFcol = "state_phase",
  statecol = "state",
  returnlist = FALSE,
  raster = FALSE,
  y_axis_trans = "identity",
  xaxis_order = "genome_position",
  legend.position = "bottom",
  genes = NULL,
  annotateregions = NULL,
  homolog = FALSE,
  SV = NULL,
  adj = 0.03,
  svalpha = 0.5,
  svwidth = 1,
  ...
)
}
\arguments{
\item{cn}{Single cell allele specific copy number dataframe with the following columns: `cell_id`, `chr`, `start`, `end`, `state`, `copy` or a hscn object.}

\item{cellid}{Which cell to plot, if no cell is specific will plot the first cell in the dataframe}

\item{chrfilt}{Vector of chromosomes to plot, if NULL (default) will plot all chromosomes}

\item{pointsize}{The point size in the plot}

\item{alphaval}{Alpha value of points}

\item{maxCN}{The maximum on the y axis, if any points are above this value they will be winsorized rather than removed}

\item{cellidx}{idx of cell to plot if cellid = NULL}

\item{statecol}{The colour mapping, default is to map colours to the `state` column}

\item{returnlist}{Return a list rather than the ggplot object}

\item{raster}{use ggrastr or not, default = FALSE}

\item{y_axis_trans}{What transformation to use on the y-axis, default is identity, the other option is "squashy" which uses a tanh transformation}

\item{xaxis_order}{Default is "genome_position"}

\item{legend.position}{Where to place the legend, default is "bottom"}

\item{genes}{vector of genes to annotate, will add a dashed vertical line and label}

\item{annotateregions}{Dataframe with chr start and end positions to annotate, will draw a dashed vertical line at this position}

\item{homolog}{Rather than plot the BAF and CN seperately this will plot the 2 homologs on the same track}

\item{SV}{Default is NULL. If a dataframe with structural variant position is passed it will add a track on the top showin rearrangement links}

\item{svalpha}{the alpha scaling of the SV lines, default = 0.5}

\item{svwidth}{width of rearrangement connections}
}
\value{
ggplot2 plot
}
\description{
Plot a single cell allele specific copy number profile
}
\examples{

\dontrun{
data("haplotypes")
data("CNbins")
haplotypes <- format_haplotypes_dlp(haplotypes, CNbins)
hscn <- callHaplotypeSpecificCN(CNbins, haplotypes, likelihood = "binomial")
plotCNprofileBAF(hscn, genes = "MYC")
plotCNprofileBAF(hscn, homolog = TRUE, chrfilt = c("1", "8"))
}

}
