% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_palettes.R
\name{scale_color_scCN}
\alias{scale_color_scCN}
\title{Colour scale constructor for scCN (single cell Copy Number) colors}
\usage{
scale_color_scCN(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in scCN_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour scale constructor for scCN (single cell Copy Number) colors
}
