% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callASCN.R
\name{callAlleleSpecificCN}
\alias{callAlleleSpecificCN}
\title{Call allele specific copy number in single cell datasets}
\usage{
callAlleleSpecificCN(
  CNbins,
  haplotypes,
  eps = 0.000000000001,
  loherror = 0.02,
  maxCN = NULL,
  selftransitionprob = 0.95,
  progressbar = TRUE,
  ncores = 1,
  likelihood = "binomial",
  minbins = 100,
  minbinschr = 10,
  maxloherror = 0.03,
  filterhaplotypes = 0.1
)
}
\arguments{
\item{CNbins}{single cell copy number dataframe with the following columns: `cell_id`, `chr`, `start`, `end`, `state`, `copy`}

\item{haplotypes}{single cell haplotypes dataframe with the following columns: `cell_id`, `chr`, `start`, `end`, `hap_label`, `allele1`, `allele0`, `totalcounts`}

\item{eps}{default 1e-12}

\item{loherror}{LOH error rate for initial assignment, this is inferred directly from the data in the second pass, default = 0.02}

\item{maxCN}{maximum copy number to infer allele specific states, default=NULL which will use the maximum state from CNbins}

\item{selftransitionprob}{probability to stay in the same state in the HMM, default = 0.999, set to 0.0 for an IID model}

\item{progressbar}{Boolean to display progressbar or not, default = TRUE, will only show if ncores == 1}

\item{ncores}{Number of cores to use, default = 1}

\item{likelihood}{Likelihood model for HMM, default is `binomial`, other option is `betabinomial` or use `auto` and the algorithm will choose the likelihood that best fits the data.}

\item{minbins}{Minimum number of bins containing both haplotype counts and copy number data for a cell to be included}

\item{minbinschr}{Minimum number of bins containing both haplotype counts and copy number data per chromosome for a cell to be included}

\item{maxloherror}{Maximum value for LOH error rate}

\item{filterhaplotypes}{filter out haplotypes present in less than X fraction, default is 0.1}
}
\value{
allele specific copy number object which includes dataframe similar to input with additional columns which include

* `Maj` (Major allele copy number)
* `Min` (Minor allele copy number)
* `state_AS_phased` (phased state of the form Maj|Min )
* `state_AS` (mirrored state of the form Maj|Min)
* `LOH` (is bin LOH or not)
* `state_phase` (state describing which is the dominant allele and whether it is LOH or not)
* `state_BAF` (binned discretized BAF value calculated as Min / (Maj + Min))
}
\description{
Call allele specific copy number in single cell datasets
}
\details{
In the allele specific copy number inference Maj is always > Min and state_AS_phased == state_AS
}
