% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clusters_used_for_phasing}
\alias{plot_clusters_used_for_phasing}
\title{Plot clusters used for phasing}
\usage{
plot_clusters_used_for_phasing(hscn)
}
\arguments{
\item{hscn}{An object of class \code{hscn} generated by \code{\link{hscn}}}
}
\value{
A \code{ggplot2} object
}
\description{
Plots the consensus copy number and BAF profiles of clusters used for phasing
}
\examples{

plot_clusters_used_for_phasing(hscn)

}
