% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCNprofile}
\alias{plotCNprofile}
\title{Plot a single cell copy number profile}
\usage{
plotCNprofile(
  CNbins,
  cellid = NULL,
  chrfilt = NULL,
  pointsize = 1,
  alphaval = 0.6,
  maxCN = 10,
  cellidx = 1,
  statecol = "state",
  returnlist = FALSE,
  raster = FALSE,
  genome = "hg19",
  y_axis_trans = "identity",
  xaxis_order = "genome_position",
  legend.position = "bottom",
  annotateregions = NULL,
  annotateregions_linetype = 2,
  SV = NULL,
  SVcol = TRUE,
  svalpha = 0.5,
  svwidth = 1,
  adj = 0.03,
  genes = NULL,
  tickwidth = 50,
  chrstart = NULL,
  chrend = NULL,
  shape = 16,
  positionticks = FALSE,
  ideogram = FALSE,
  overwrite_color = NULL,
  sv_style = "curves",
  sv_arc_alpha = 0.3,
  sv_arc_height = 0.2,
  show_sv_read_axis = TRUE,
  sv_read_axis_scale = NULL,
  show_sv_legend = FALSE,
  regions = NULL,
  region_gap = 5,
  ...
)
}
\arguments{
\item{CNbins}{Single cell copy number dataframe with the following columns: \code{cell_id}, \code{chr}, \code{start}, \code{end}, \code{state}, \code{copy}}

\item{cellid}{Which cell to plot, if no cell is specific will plot the first cell in the dataframe}

\item{chrfilt}{Vector of chromosomes to plot, if NULL (default) will plot all chromosomes}

\item{pointsize}{The point size in the plot}

\item{alphaval}{Alpha value of points}

\item{maxCN}{The maximum on the y axis, if any points are above this value they will be winsorized rather than removed}

\item{cellidx}{idx of cell to plot if cellid = NULL}

\item{statecol}{The colour mapping, default is to map colours to the \code{state} column}

\item{returnlist}{Return a list rather than the ggplot object}

\item{raster}{use ggrastr or not, default = FALSE}

\item{genome}{genome to use, default = "hg19" (only used for ideogram)}

\item{y_axis_trans}{What transformation to use on the y-axis, default is identity, the other option is "squashy" which uses a tanh transformation}

\item{xaxis_order}{Default is "genome_position"}

\item{legend.position}{Where to place the legend, default is "bottom"}

\item{annotateregions}{Dataframe with chr start and end positions to annotate, will draw a dashed vertical line at this position}

\item{annotateregions_linetype}{linetype for region annotation, default = 2 (dashed)}

\item{SV}{Default is NULL. If a dataframe with structural variant position is passed it will add rearrangement links between bins. For lines_and_arcs style, SV must include columns: strand_1, strand_2 (values '+' or '-'), and read_count.}

\item{SVcol}{Default is TRUE. Colour SVs or not}

\item{svalpha}{the alpha scaling of the SV lines, default = 0.5}

\item{svwidth}{Width of SV width curves, default = 1.0}

\item{adj}{adjustment for gene labels}

\item{genes}{vector of genes to annotate, will add a dashed vertical line and label}

\item{tickwidth}{Spacing of ticks (in Mb) when only 1 chromosome is plotted}

\item{chrstart}{Start of region (in Mb) when plotting a single chromosome}

\item{chrend}{End of region (in Mb) when plotting a single chromosome}

\item{shape}{shape for plotting, default = 16}

\item{positionticks}{set to TRUE to use position ticks rather than chromosome ticks}

\item{ideogram}{plot ideogram at the top, default = TRUE}

\item{sv_style}{visualization style for structural variants: "curves" (default, bezier curves), "lines_and_arcs" (colored lines with arcs), or "both"}

\item{sv_arc_alpha}{transparency of SV arcs when using lines_and_arcs style, default = 0.3}

\item{sv_arc_height}{height of SV arcs as fraction of maxCN when using lines_and_arcs style, default = 0.2}

\item{show_sv_read_axis}{show secondary y-axis for SV read support when using lines_and_arcs style, default = TRUE. Works with both identity and squashy y-axis transformations.}

\item{sv_read_axis_scale}{maximum value for SV read support axis (auto-scaled if NULL), default = NULL}

\item{show_sv_legend}{show legend for SV orientations when using lines_and_arcs style, default = FALSE}

\item{regions}{Optional data.frame with columns \code{chr}, \code{start}, \code{end} (in Mb) to plot multiple regions. If provided, takes precedence over \code{chrfilt}/\code{chrstart}/\code{chrend}. The order of rows determines the order of regions in the plot.}

\item{region_gap}{Gap size between regions in Mb, default = 5. Only used when \code{regions} is provided.}
}
\value{
ggplot2 plot
}
\description{
Plot a single cell copy number profile
}
\examples{

plotCNprofile(CNbins)

# Plot multiple regions:
# regions <- data.frame(
#   chr = c("2", "2", "1"),
#   start = c(1, 100, 50),
#   end = c(10, 120, 75)
# )
# plotCNprofile(CNbins, regions = regions, region_gap = 5)
}
