% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap plot}
\usage{
plotHeatmap(
  cn,
  tree = NULL,
  clusters = NULL,
  normalize_ploidy = FALSE,
  normalize_tree = FALSE,
  branch_length = 1,
  spacer_cols = 20,
  plottree = TRUE,
  plotcol = "state",
  reorderclusters = FALSE,
  pctcells = 0.05,
  library_mapping = NULL,
  clone_pal = NULL,
  sample_label_idx = 1,
  fillna = TRUE,
  frequencycutoff = 2,
  maxf = NULL,
  plotfrequency = FALSE,
  show_legend = TRUE,
  show_library_label = TRUE,
  show_clone_label = TRUE,
  widenarm = FALSE,
  umapmetric = "euclidean",
  chrlabels = TRUE,
  raster_quality = 10,
  SV = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{cn}{Either a hscn object or a single cell allele specific copy number dataframe with the following columns: `cell_id`, `chr`, `start`, `end`, `state`, `copy`}

\item{tree}{Tree in newick format to plot alongside the heatmap, default = NULL}

\item{normalize_ploidy}{Normalize ploidy of all cells to 2}

\item{normalize_tree}{default = FALSE}

\item{branch_length}{scales branch lengths to this size, default = 2}

\item{spacer_cols}{number of empty columns between chromosomes, default = 20}

\item{plottree}{Binary value of whether to plot tree or not, default = TRUE}

\item{plotcol}{Which column to colour the heatmap by, should be one of "state", "state_BAF", "state_phase", "state_AS", "state_min", "copy", "BAF", "Min", "Maj"}

\item{pctcells}{Minimum size of cluster in terms of perecentage of cells in umap clustering}

\item{library_mapping}{Named vector mapping library names to labels for legend}

\item{clone_pal}{pallette to colour clusters by}

\item{sample_label_idx}{default = 1}

\item{fillna}{Smooth over NA values, default = TRUE}

\item{frequencycutoff}{default = 2}

\item{maxf}{Max frequency when plotting the frequency track, default = NULL infers this from the data}

\item{plotfrequency}{Plot the frequency track of gains and losses across the genome}

\item{show_legend}{plot legend or not, boolean}

\item{show_library_label}{show library label or not, boolean}

\item{show_clone_label}{show clone label or not, boolean}

\item{umapmetric}{metric to use in umap dimensionality reduction if no clusters are specified}

\item{chrlabels}{include chromosome labels or not, boolean}

\item{SV}{sv data frame}

\item{seed}{seed for UMAP

If clusters are set to NULL then the function will compute clusters using UMAP and HDBSCAN.

#' \dontrun{
data("haplotypes")
data("CNbins")
haplotypes <- format_haplotypes_dlp(haplotypes, CNbins)
hscn <- callHaplotypeSpecificCN(CNbins, haplotypes, likelihood = "binomial")
plotHeatmap(hscn)
}}

\item{cluster}{data.frame assigning cells to clusters, needs the following columns `cell_id`, `clone_id` default = NULL}

\item{reordercluster}{Reorder the cells according to cluster if no tree is specified}
}
\description{
Plot a heatmap where rows are cells, columns are genome coordinates and colours map to (allele-specific) copy-number states
}
