% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap plot}
\usage{
plotHeatmap(
  cn,
  tree = NULL,
  clusters = NULL,
  normalize_ploidy = FALSE,
  normalize_tree = FALSE,
  branch_length = 1,
  spacer_cols = 20,
  plottree = TRUE,
  plotcol = "state",
  reorderclusters = FALSE,
  pctcells = 0.05,
  library_mapping = NULL,
  clone_pal = NULL,
  sample_label_idx = 1,
  fillna = TRUE,
  frequencycutoff = 2,
  maxf = NULL,
  plotfrequency = FALSE,
  frequency_height = 1.4,
  show_legend = TRUE,
  show_library_label = TRUE,
  show_clone_label = TRUE,
  show_clone_text = TRUE,
  widenarm = FALSE,
  umapmetric = "euclidean",
  chrlabels = TRUE,
  labeladjust = -5,
  SV = NULL,
  seed = NULL,
  nticks = 4,
  Mb = TRUE,
  fillgenome = FALSE,
  annotation_height = NULL,
  annofontsize = 10,
  na_col = "white",
  linkheight = 2.5,
  newlegendname = NULL,
  str_to_remove = NULL,
  maxCNcol = 11,
  anno_width = 0.4,
  rasterquality = 15,
  tree_width = 4,
  ...
)
}
\arguments{
\item{cn}{Either a hscn object or a single cell allele specific copy number dataframe with the following columns: `cell_id`, `chr`, `start`, `end`, `state`, `copy`}

\item{tree}{Tree in newick format to plot alongside the heatmap, default = NULL}

\item{clusters}{data.frame assigning cells to clusters, needs the following columns `cell_id`, `clone_id` default = NULL}

\item{normalize_ploidy}{Normalize ploidy of all cells to 2}

\item{normalize_tree}{default = FALSE}

\item{branch_length}{scales branch lengths to this size, default = 2}

\item{spacer_cols}{number of empty columns between chromosomes, default = 20}

\item{plottree}{Binary value of whether to plot tree or not, default = TRUE}

\item{plotcol}{Which column to colour the heatmap by, should be one of "state", "state_BAF", "state_phase", "state_AS", "state_min", "copy", "BAF", "A", "B"}

\item{reorderclusters}{Reorder the cells according to cluster if no tree is specified}

\item{pctcells}{Minimum size of cluster in terms of perecentage of cells in umap clustering}

\item{library_mapping}{Named vector mapping library names to labels for legend}

\item{clone_pal}{pallette to colour clusters by}

\item{sample_label_idx}{default = 1}

\item{fillna}{Smooth over NA values, default = TRUE}

\item{frequencycutoff}{default = 2}

\item{maxf}{Max frequency when plotting the frequency track, default = NULL infers this from the data}

\item{plotfrequency}{Plot the frequency track of gains and losses across the genome}

\item{frequency_height}{height of the frequency track if using, default = 1.4}

\item{show_legend}{plot legend or not, boolean}

\item{show_library_label}{show library label or not, boolean}

\item{show_clone_label}{show clone label or not, boolean}

\item{show_clone_text}{Show small inset labels next to clone/cluster annotation}

\item{widenarm}{Widen the copy number data table to include all bins}

\item{umapmetric}{metric to use in umap dimensionality reduction if no clusters are specified}

\item{chrlabels}{include chromosome labels or not, boolean}

\item{labeladjust}{}

\item{SV}{sv data frame}

\item{seed}{seed for UMAP}

\item{nticks}{number of ticks in x-axis label when plotting a single chromosome}

\item{Mb}{Use Mb ticks when plotting single chromosome}

\item{fillgenome}{fill in any missing bins and add NA to centromeric regions}

\item{annotation_height}{Height of the annotations}

\item{annofontsize}{Font size to use for annotations, default = 10}

\item{na_col}{colour of NA values}

\item{linkheight}{height of x-axis ticks}

\item{newlegendname}{overwrite default legend name}

\item{str_to_remove}{string to remove from cell_id's when plotting labels}

\item{maxCNcol}{max value for color scale when plotting raw data}

\item{anno_width}{width of left annotations}

\item{rasterquality}{default = 15}

\item{tree_width}{Width of phylogenetic tree, default = 4

If clusters are set to NULL then the function will compute clusters using UMAP and HDBSCAN.}
}
\description{
Plot a heatmap where rows are cells, columns are genome coordinates and colours map to (allele-specific) copy-number states
}
\examples{
\dontrun{
data("haplotypes")
data("CNbins")
haplotypes <- format_haplotypes_dlp(haplotypes, CNbins)
hscn <- callHaplotypeSpecificCN(CNbins, haplotypes, likelihood = "binomial")
plotHeatmap(hscn)
}

}
