% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callHSCN.R
\name{rephasebins}
\alias{rephasebins}
\title{Rephase haplotype copy number}
\usage{
rephasebins(cn, chromosomes = NULL)
}
\arguments{
\item{cn}{either a \code{hscn} object from \code{callHaplotypeSpecificCN} or a dataframe with haplotype specific copy number ie the \code{data} slot in an \code{hscn} object}

\item{chromosomes}{vector specifying which chromosomes to phase, default is NULL whereby all chromosomes are phased}
}
\value{
Either a new \code{hscn} object or a dataframe with rephased bins depdending on the input
}
\description{
This function implements the dynamic programming algorithm to rephase haplotype copy number first described in CHISEL.
The objective is to find the phase that minimizes the number of copy number events
}
